% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm.R
\name{MixtureNormal}
\alias{MixtureNormal}
\title{The Mixture Normal Distribution}
\usage{
MixtureNormal(p, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{p}{is the probability of sampling from the first normal distribution}

\item{mu1}{and @param sigma1 are the parameters of this first distribution,
and @param mu2 / @param sigma2 are the parameters of the second distribution,
which has a probability of being sampled of 1-p.}
}
\description{
This function allows you to sample from a mixture normal distribution with
user-inputted probabilities, means, and standard deviations
}
\examples{
MixtureNormal(0.5, 5, 5, 1, 10)
}
